backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
groq
motor
passlib[bcrypt]
bcrypt==4.1.3
python-jose[cryptography]
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
pydantic==2.9.2
email-validator==2.1.1
langgraph>=1.0.0,<2
langchain-groq>=1.0.0,<2
langchain-core>=1.0.0,<2
chromadb>=0.5.0,<1
sentence-transformers>=3.0.0,<4
pytest==8.3.4
